completion: handle dynamic completions, and completions
The first part of this patch is to check if the autocomplete choices is
a function, and if so invoke it to return a list that can then be used
to populate the autocompletion suggestions.

The second part of this patch is to fix the bug that exists for all
value completions for a keyword. Without this patch, the only way to
select a choice is by using the up/down arrow keys; you cannot first
trim the selection by typing a few letters and then selecting.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>