Limit completions offered when completions are too many
In case the number of completions is too many, say the list of hostnames
in a 10000 node network, instead of trying to show all 10K hostnames as
completions, do it only for the first ten. But the entire list exists,
and so you can still start typing to get completions for the ones not
listed yet.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>